Remove unused util file, apply HPE theme
